pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "agentstress"
version = "0.1.0"
description = "Multi-agent reliability testing framework — Jepsen/Chaos Mesh for AI agent teams"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "wjeong"}]

dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "pydantic>=2.6.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "rich>=13.7.0",
    "tiktoken>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]
judge = [
    "vllm>=0.4.0",
    "openai>=1.12.0",
]

[project.scripts]
agentstress = "cli.agentstress_cli_main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["test"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120
target-version = "py310"
